Split schemas for project and dataset
